Conflicts:
	src/engine/galaxy/ep5/ai/CShikadiMine.h

+++ src/engine/galaxy/ep5/ai/CShelly.cpp
@@ -12,59 +12,60 @@
 #include "misc.h"
 
 /*
-$21DCW #Lil Ampton walk
-$21FAW #Lil Ampton walk
-$2218W #Lil Ampton walk
-$2236W #Lil Ampton walk
-$2254W #Lil Ampton turn 4
-$2272W #Lil Ampton start pole slide 5
-$2290W #Lil Ampton start pole slide
-$22AEW #Lil Ampton pole slide 6
-$22CCW #Lil Ampton stop pole slide 7
-$22EAW #Lil Ampton stop pole slide
-$2308W #Lil Ampton flip switch 8
-$2326W #Lil Ampton flip switch 
-$2344W #Lil Ampton flip switch
-$2362W #Lil Ampton flip switch
-$2380W #Lil Ampton flip switch
-$239EW #Stunned Ampton 13
+$2416W #Shelly walk
+$2434W #Shelly walk
+$2452W #Shelly walk
+$2470W #Shelly walk
+$248EW #Shelly prepare to jump 4
+$24ACW #Shelly jump 5
+$24CAW #Shelly fall 6
+$24E8W #Shelly fall
+$2506W #Shelly smash 8
+$2524W #Shelly smash
+$2542W #Shelly smash smoke 10
+$2560W #Shelly smash smoke
+$257EW #Shelly smash smoke
+$259CW #Shelly smash smoke
+$25BAW #Shelly fragments 14
+$25D8W #Shelly fragments 15 -> TODO: Need to be other actions!
  */
 
 
 namespace galaxy {  
   
-enum SPARKYACTIONS
+enum SHELLYACTIONS
 {
-A_AMPTON_WALK = 0,
-A_AMPTON_TURN = 4,
-A_AMPTON_START_POLE = 5,
-A_AMPTON_POLE_SLIDE = 6,
-A_AMPTON_STOP_POLE = 7,
-A_AMPTON_FLIP_SWITCH = 8,
-A_AMPTON_STUNNED = 12
+A_SHELLY_WALK = 0,
+A_SHELLY_PREPARE_JUMP = 4,
+A_SHELLY_JUMP = 5,
+A_SHELLY_FALL = 6,
+A_SHELLY_SMASH = 8,
+A_SHELLY_SMOKE = 10
 };
 
-const int TIME_UNTIL_MOVE = 5;
-const int TIME_FOR_LOOK = 150;
+const int WALK_SPEED = 10;
 
-const int WALK_SPEED = 25;
+const int JUMP_X_SPEED = 50;
 
-const int CSF_DISTANCE_TO_FOLLOW = 6<<CSF;
+const int SMOKE_TIME = 40;
 
-const int CHARGE_TIME = 250;
-const int CHARGE_SPEED = 75;
+const int MAX_JUMP_INERTIA = 150;
 
-const int TURN_TIME = 10;
-
+const int CSF_DISTANCE_TO_JUMP = 6<<CSF;
   
 CShelly::CShelly(CMap *pmap, const Uint16 foeID, const Uint32 x, const Uint32 y) :
 CStunnable(pmap, foeID, x, y),
 mTimer(0)
 {
-	mActionMap[A_AMPTON_STUNNED] = &CStunnable::processGettingStunned;
+      	mActionMap[A_SHELLY_WALK] = (void (CStunnable::*)()) &CShelly::processWalking;
+      	mActionMap[A_SHELLY_PREPARE_JUMP] = (void (CStunnable::*)()) &CShelly::processPrepareJump;
+      	mActionMap[A_SHELLY_JUMP] = (void (CStunnable::*)()) &CShelly::processJump;
+      	mActionMap[A_SHELLY_FALL] = (void (CStunnable::*)()) &CShelly::processFall;
+      	mActionMap[A_SHELLY_SMASH] = (void (CStunnable::*)()) &CShelly::processSmash;
+      	mActionMap[A_SHELLY_SMOKE] = (void (CStunnable::*)()) &CShelly::processSmoke;
   
 	// Adapt this AI
-	setupGalaxyObjectOnMap(0x21DC, A_AMPTON_WALK);
+	setupGalaxyObjectOnMap(0x2416, A_SHELLY_WALK);
 	
 	xDirection = LEFT;
 }
@@ -72,32 +73,125 @@ mTimer(0)
 
 
 void CShelly::processWalking()
+{ 
+    
+    if( blockedl )
+    {
+	xDirection = RIGHT;
+    }
+    else if(blockedr)
+    {
+	xDirection = LEFT;
+    }
+    
+    // Move normally in the direction
+    moveXDir(xDirection*WALK_SPEED);   
+        
+    if(mGoodJumpChance)
+    {
+	setAction(A_SHELLY_PREPARE_JUMP);
+    }
+}
+
+
+void CShelly::processPrepareJump()
 {
-  // Move normally in the direction
-  if( xDirection == RIGHT )
-  {
-    moveRight( WALK_SPEED );
-  }
-  else
-  {
-    moveLeft( WALK_SPEED );
-  }   
+    if(getActionStatus(A_SHELLY_JUMP))
+    {
+	if(xDirection != mKeenAlignment)
+	{
+	    xDirection = mKeenAlignment;
+	    setAction(A_SHELLY_WALK);
+	    return;
+	}
+	
+	if(mGoodJumpChance)
+	{
+	    setAction(A_SHELLY_JUMP);
+	    yinertia = -MAX_JUMP_INERTIA;
+	}
+	else
+	{
+	    setAction(A_SHELLY_WALK);
+	}
+    }
+}
+
+
+void CShelly::processJump()
+{
+    if( getActionStatus(A_SHELLY_WALK) )
+    {
+	setAction(A_SHELLY_FALL);
+    }
+    
+    if( getActionStatus(A_SHELLY_FALL)  )
+    {
+	setAction(A_SHELLY_FALL);
+    }
+}
+
+
+void CShelly::processFall()
+{
+
+    moveXDir(xDirection*JUMP_X_SPEED);
+    
+    if(blockedd)
+	setAction(A_SHELLY_SMASH);
+}
+
+
+void CShelly::processSmash()
+{
+    if( blockedd )
+	setAction(A_SHELLY_SMOKE);
+}
+
+
+void CShelly::processSmoke()
+{
+    mTimer++;
+    
+    if(mTimer < SMOKE_TIME)
+	return;
+    
+    mTimer = 0;
+    
+    // TODO: Here we need to spawn deadly shards, we should base this on the Shards Class used in Keen 4
+    
+    exists = false;
 }
 
 
 bool CShelly::isNearby(CSpriteObject &theObject)
 {
-	if( !getProbability(10) )
-		return false;
-
 	if( CPlayerLevel *player = dynamic_cast<CPlayerLevel*>(&theObject) )
 	{
-		/*if( player->getXMidPos() < getXMidPos() )
+		if( player->getXMidPos() < getXMidPos() )
 			mKeenAlignment = LEFT;
 		else
-			mKeenAlignment = RIGHT;*/
-	}
+			mKeenAlignment = RIGHT;
+		
+		
+		const int objX = theObject.getXMidPos();
+		const int objY = theObject.getYMidPos();
+		const int shellyX = getXMidPos();
+		const int shellyY = getYMidPos();
+		
+		mGoodJumpChance = false;
+		
+		if( objX < shellyX - CSF_DISTANCE_TO_JUMP ||
+			objX > shellyX + CSF_DISTANCE_TO_JUMP )
+			return false;
 
+		if( objY < shellyY - CSF_DISTANCE_TO_JUMP ||
+			objY > shellyY + CSF_DISTANCE_TO_JUMP )
+			return false;
+		
+		mGoodJumpChance = true;
+	}
+	
 	return true;
 }
 
@@ -111,15 +205,13 @@ void CShelly::getTouchedBy(CSpriteObject &theObject)
 	// Was it a bullet? Than make it stunned.
 	if( dynamic_cast<CBullet*>(&theObject) )
 	{
-		playSound(SOUND_ROBO_STUN);
-		dead = true;
 		theObject.dead = true;
 	}
 
-	/*if( CPlayerBase *player = dynamic_cast<CPlayerBase*>(&theObject) )
+	if( CPlayerBase *player = dynamic_cast<CPlayerBase*>(&theObject) )
 	{
 		player->kill();
-	}*/
+	}
 }
 
 
@@ -135,19 +227,9 @@ void CShelly::process()
 {
 	performCollisions();
 	
-	performGravityMid();
+	performGravityHigh();
 
-	if( blockedl )
-	{
-	  xDirection = RIGHT;
-	}
-	else if(blockedr)
-	{
-	  xDirection = LEFT;
-	}
-
-	if(!processActionRoutine())
-	    exists = false;
+	processActionRoutine();
 	
 	(this->*mp_processState)();
 }
